# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Report Studio API
  version: "2.0"
  description: |
    Report Studio backend HTTP API (Javalin).

    The API is split by resource, not by a single version prefix: `/api/v2/*`
    carries the current ReportDefinition-based stack (templates and everything
    under them, schemas, pdf/pdf-jobs, scalardb, documents, excel, tenant,
    health), while `/api/v1/*` is the permanent peripheral-platform namespace
    (auth/PAT, admin, products, jobs, webhooks, sequences, public forms,
    binding-trees, health). The v1 namespace is frozen — new resources always
    go to v2 — and only `jobs` is slated for eventual consolidation into
    `/api/v2/pdf-jobs` (policy decided in #278). The version lives in the URL,
    not in a class name.

    **Authoritative route registration:** `server/src/main/java/com/report/server/ApiRoutes.java`.
    A parity test (`OpenApiRouteParityTest`) fails CI if this document and that
    file diverge, so every registered route is documented here and vice-versa.

    ## Authentication
    Two mechanisms, resolved by the `/api/*` before-filter:
    - **cookieAuth** — browser session cookie issued by `POST /api/v1/auth/login`.
    - **bearerAuth** — Personal Access Token (`rpat_…`) in `Authorization: Bearer`
      (#195), the official path for CLI / CI / external programmatic use.

    Public endpoints (`/api/v1/public/*`, `/api/v1/auth/login|logout|me`, and both
...
    (Origin/Referer); Bearer requests are exempt.
servers:
...
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Browser session cookie from POST /api/v1/auth/login.
    bearerAuth:
      type: http
      scheme: bearer
      description: Personal Access Token (rpat_…), for CLI / CI / external clients (#195).
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Password login; sets the session cookie
      security: []
      responses:
        '200': { description: Logged in }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/auth/change-profile:
    post:
      tags: [auth]
      summary: Change display name / password (invalidates other sessions)
      responses:
        '200': { description: Updated }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ── Personal Access Tokens (#195) ──────────────────────────────────────────
  /api/v1/auth/tokens:
    get:
      tags: [tokens]
      summary: List the caller's PATs (metadata only)
      responses:
        '200': { description: Token list }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [tokens]
      summary: Create a PAT (plaintext returned once)
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/tokens/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    delete:
      tags: [tokens]
      summary: Revoke a PAT
      responses:
        '200': { description: Revoked }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Admin (requires admin role) ─────────────────────────────────────────────
...
  /api/v1/admin/users/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [admin]
      summary: Update a user
      responses: { '200': { description: Updated } }
    delete:
      tags: [admin]
      summary: Delete a user
      responses: { '200': { description: Deleted } }
...
  /api/v1/binding-trees/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [binding-trees]
      summary: Get a binding tree
      responses:
        '200': { description: Binding tree }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [binding-trees]
      summary: Upsert a binding tree
      responses: { '200': { description: Saved } }

  # ── Products ────────────────────────────────────────────────────────────────
...
  /api/v1/products/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [products]
      summary: Get a product
      responses:
        '200': { description: Product }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [products]
      summary: Update a product (optimistic concurrency)
      responses: { '200': { description: Updated } }
    delete:
      tags: [products]
      summary: Delete a product
      responses: { '200': { description: Deleted } }
...
  /api/v1/jobs/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [jobs]
      summary: Get job status
      responses:
        '200': { description: Job }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [jobs]
      summary: Cancel / delete a job
      responses: { '200': { description: Deleted } }
  /api/v1/jobs/{id}/output:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [jobs]
      summary: Download a job's output artifact
      responses:
        '200': { description: Artifact (ZIP/PDF) }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── Public forms (unauthenticated) ──────────────────────────────────────────
  /api/v1/public/forms/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [public-forms]
      summary: Public form info (published + password requirement)
      security: []
      responses:
        '200': { description: Form info }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/public/forms/{id}/projection:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [public-forms]
      summary: Public form projection (session required if password-protected)
      security: []
      responses:
        '200': { description: Projection (schemaGroups stripped) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/public/forms/{id}/verify:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [public-forms]
      summary: Verify form password (rate-limited); sets form_session cookie
      security: []
      responses:
        '200': { description: Verified }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/public/forms/{id}/submit:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [public-forms]
      summary: Submit a public form response
      security: []
      responses:
        '201': { description: Created }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ── Sequences ───────────────────────────────────────────────────────────────
  /api/v1/sequences/{templateId}:
    parameters: [{ $ref: '#/components/parameters/TemplateId' }]
    get:
      tags: [sequences]
      summary: Get the document-number sequence config
      responses: { '200': { description: Sequence } }
    put:
      tags: [sequences]
      summary: Update the sequence config
      responses: { '200': { description: Updated } }

  # ── Webhooks ────────────────────────────────────────────────────────────────
  /api/v1/webhooks/{templateId}:
    parameters: [{ $ref: '#/components/parameters/TemplateId' }]
    get:
      tags: [webhooks]
      summary: Get the webhook config
      responses: { '200': { description: Webhook } }
    put:
      tags: [webhooks]
      summary: Update the webhook config
      responses: { '200': { description: Updated } }
  /api/v1/webhooks/{templateId}/test:
    parameters: [{ $ref: '#/components/parameters/TemplateId' }]
    post:
      tags: [webhooks]
      summary: Send a test webhook (SSRF-guarded)
      responses: { '200': { description: Dispatched } }

  # ── Templates ───────────────────────────────────────────────────────────────
...
  /api/v2/templates/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [templates]
      summary: Get a template
      responses:
        '200':
          description: Template (canonical envelope)
          content:
            application/json:
              schema:
                type: object
                required: [formatVersion, id, name, createdAt, updatedAt, visibility, definition]
                properties:
                  formatVersion: { type: integer }
                  id: { type: string }
                  name: { type: string }
                  createdAt: { type: string }
                  updatedAt: { type: string }
                  visibility: { type: string }
                  definition: { type: object }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [templates]
      summary: Replace a template
      responses: { '200': { description: Updated } }
    post:
      tags: [templates]
      summary: Update a template (POST alias)
      responses: { '200': { description: Updated } }
    delete:
      tags: [templates]
      summary: Delete a template
      responses: { '200': { description: Deleted } }
  /api/v2/templates/{id}/export:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [templates]
      summary: Export a template envelope (.rds2.json)
      responses: { '200': { description: Envelope } }
  /api/v2/templates/{id}/thumbnail:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [templates]
      summary: Template thumbnail (PNG)
      responses: { '200': { description: PNG } }
  /api/v2/templates/{id}/copy:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [templates]
      summary: Copy a template
      responses: { '201': { description: Copied } }
  /api/v2/templates/{id}/duplicate:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [templates]
      summary: Duplicate a template
      responses: { '201': { description: Duplicated } }
  /api/v2/templates/{id}/visibility:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [templates]
      summary: Change template visibility (public/private)
      responses: { '200': { description: Updated } }
  /api/v2/templates/{id}/evaluate:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [templates]
      summary: Evaluate calculations for a data payload
      responses: { '200': { description: Evaluation } }
  /api/v2/templates/{id}/validate:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [templates]
      summary: Validate a data payload against the template rules
      responses:
        '200': { description: Valid }
        '422': { description: Validation failed }
  /api/v2/templates/{id}/resolve-bindings:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [templates]
      summary: Resolve schema bindings against ScalarDB
      responses:
        '200': { description: Resolved }
        '207': { description: Partial success }
  /api/v2/templates/{id}/pdf:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [pdf]
      summary: Render a template to PDF (synchronous)
      responses: { '200': { description: PDF } }

  # ── Template responses ──────────────────────────────────────────────────────
  /api/v2/templates/{id}/responses:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [template-responses]
      summary: List responses for a template
      responses:
        '200':
          description: Responses (paginated summary items)
          content:
            application/json:
              schema:
                type: object
                required: [items, total, offset, limit, hasMore]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, templateId, submittedAt, status, summary]
                      properties:
                        id: { type: string }
                        templateId: { type: string }
                        submittedAt: { type: integer }
                        submittedBy: { type: string }
                        status: { type: string }
                        summary:
                          type: array
                          items: { type: string }
                  total: { type: integer }
                  offset: { type: integer }
                  limit: { type: integer }
                  hasMore: { type: boolean }
    post:
      tags: [template-responses]
      summary: Create a response (may stamp a document number)
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [id, templateId, submittedAt]
                properties:
                  id: { type: string }
                  templateId: { type: string }
                  submittedAt: { type: integer }
  /api/v2/templates/{id}/responses/export:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [template-responses]
      summary: Export responses (CSV/JSON)
      responses: { '200': { description: Export } }
  /api/v2/templates/{id}/responses/{rid}:
    parameters:
      - { $ref: '#/components/parameters/Id' }
      - { $ref: '#/components/parameters/Rid' }
    get:
      tags: [template-responses]
      summary: Get a response
      responses:
        '200':
          description: Response (stored document, raw)
          content:
            application/json:
              schema:
                type: object
                required: [id, templateId, data]
                properties:
                  id: { type: string }
                  templateId: { type: string }
                  data: { type: object }
                  submittedAt: { type: integer }
                  submittedBy: { type: string }
                  status: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [template-responses]
      summary: Delete a response
      responses: { '200': { description: Deleted } }
...